Rename sources to lowercase
