+++ src/core/containers/avltree.cpp
@@ -1,4 +1,4 @@
-#include "AVLTree.h"
+#include "avltree.h"
 #include <algorithm>
 #include <iostream>
 #include <vector>
