+++ src/core/models/fio.cpp
@@ -1,4 +1,4 @@
-#include "FIO.h"
+#include "fio.h"
 
 std::ostream& operator<<(std::ostream& os, const FIO& fio) {
     os << fio.surname << " " << fio.name << " " << fio.patronymic;
