+++ src/core/models/teacher.cpp
@@ -1,4 +1,4 @@
-#include "Teacher.h"
+#include "teacher.h"
 
 std::ostream& operator<<(std::ostream& os, const Teacher& teacher) {
     os << teacher.surname << " " << teacher.initials;
