+++ src/gui/dialogs.cpp
@@ -1,4 +1,4 @@
-#include "MainWindow.h"
+#include "mainwindow.h"
 #include <QDialog>
 #include <QFormLayout>
 #include <QLineEdit>
