+++ src/gui/mainwindow.cpp
@@ -1,4 +1,4 @@
-#include "MainWindow.h"
+#include "mainwindow.h"
 #include "ui_mainwindow.h"
 #include <QDateEdit>
 #include <QMessageBox>
@@ -17,7 +17,7 @@
 #include <fstream>
 #include <vector>
 #include <array>
-#include "FIO.h"
+#include "fio.h"
 
 MainWindow::MainWindow(HashTable* studentsTable, AVLTree* concertTree,
                        const QString& studFile, const QString& concFile,
